Move the flakiness dashboard to the new instance at webkit-test-results.appspot.com https://bugs.webkit.org/show_bug.cgi?id=114080
Reviewed by Ryosuke Niwa.

This way, the WebKit and Blink dashboard code can change as appropriate without
needing to consider the other project's needs. This still doesn't update the
buildbots to upload to the new location, or the dashboard code to pull the
results.json files from the new location, but it at least allows for changing the
UI code (e.g. the WebKit one can remove all knowledge of the Chromium bots).

* TestResultServer/app.yaml:


Identifier: 132462@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147833 268f45cc-cd09-0410-ab3c-d52691b4dbfc