REGRESSION (142152): ensure we skip past out-of-flow objects when detecting whitespace to ignore after leading empty inlines https://bugs.webkit.org/show_bug.cgi?id=114311
Reviewed by Emil A Eklund.

Source/WebCore:

When we hit an empty inline in line layout and try to detect whether it has any whitespace after it that we ought to ignore, then
we need to skip past out-of-flow and floating objects when looking for that whitespace. Failure to do so will result in us
adding the width from the first space we encounter to our line width and over-estimating the real width of the line.

We're hitting this bug now because we have broadened the category of empty inlines that get lineboxes so we go into line layout
and have to deal with them and their subsequent whitespace there. Previously this sort of whitespace would have been simply
consumed by |skipLeadingWhitespace|.

Test: fast/inline/out-of-flow-objects-and-whitespace-after-empty-inline.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::shouldSkipWhitespaceAfterStartObject):

LayoutTests:

* fast/inline/out-of-flow-objects-and-whitespace-after-empty-inline-expected.txt: Added.
* fast/inline/out-of-flow-objects-and-whitespace-after-empty-inline.html: Added.


Identifier: 132836@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148223 268f45cc-cd09-0410-ab3c-d52691b4dbfc