Fix flakiness dashboard unit test failures https://bugs.webkit.org/show_bug.cgi?id=114518
Reviewed by Ryosuke Niwa.

Fix the current failures in unit tests for the flakiness dashboard by adjusting failing
test cases to simulate loading the builders list from the webkit.org build master, replacing
the Chromium builders and platforms that are used in tests with builders and platforms for the
Apple Mac Lion WebKit2 port, and removing two test cases that were testing Chromium-specific
functionality of the dashboard.

This is sort of an interlude to the flakiness dashboard cleanup, with the removal of Chromium-specific
code and hopefully general simplification of the dashboard codebase coming in near future.

* TestResultServer/static-dashboards/flakiness_dashboard_unittests.js:
(resetGlobals):
(test):
* TestResultServer/static-dashboards/loader_unittests.js:


Identifier: 132934@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148349 268f45cc-cd09-0410-ab3c-d52691b4dbfc