Fix three crashes seen on the bots after "Make Frame's ScriptController an OwnPtr and remove the #include"
* xml/XMLHttpRequest.cpp:
(WebCore::XMLHttpRequest::open):
Check that the frame is not null (as it can be in cases like http/tests/xmlhttprequest/detaching-frame-2.html).
We used to be getting lucky, in that shouldBypassMainWorldContentSecurityPolicy(), the function that is ultimately
called, only operates on global state. Now that we need to actually dereference the Frame to get the ScriptController,
we see this crash.



Identifier: 132959@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148380 268f45cc-cd09-0410-ab3c-d52691b4dbfc