After switching to another space and back, video races to catch up with audio https://bugs.webkit.org/show_bug.cgi?id=114634
Reviewed by Tim Horton.

In Mountain Lion and previous, report that a view is "not visible" when its window
is not in the active space. This causes us to notify AVFoundation that its CAImageQueues
will not be serviced and results in not having the "catch up" behavior when returning
to the browser's space.

* UIProcess/API/mac/PageClientImpl.mm:
(WebKit::PageClientImpl::isViewVisible):

Identifier: 133019@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148464 268f45cc-cd09-0410-ab3c-d52691b4dbfc