Division by zero in CSSGradientValue::addStops() https://bugs.webkit.org/show_bug.cgi?id=114807
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2013-04-21
Reviewed by Dean Jackson.

Source/WebCore:

The normalization of the color stop positions of a linear gradient
doesn't take into account that the first and last position can be
the same. In such case the positions are computed dividing by 0.
Clamp the positions to 1 without moving the end points in such
case to match what other browsers do. This ensures that positions
passed to the platform gradient are in the 0-1 range which fixes a
crash due to an assert in BlackBerry port running test
fast/forms/type-after-focus-rule-shrink-width.html.

Test: fast/gradients/css3-color-stop-invalid.html

* css/CSSGradientValue.cpp:
(WebCore::CSSGradientValue::addStops):

LayoutTests:

* fast/gradients/css3-color-stop-invalid-expected.html: Added.
* fast/gradients/css3-color-stop-invalid.html: Added.

Identifier: 133360@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148859 268f45cc-cd09-0410-ab3c-d52691b4dbfc